requirements.txt
mysql-connector-python
python-dotenv

pip install -r requirements.txt
